+++ R/getMasterListRaw.R
@@ -7,7 +7,7 @@
 #'
 #' @param sessionID Session id integer value. Can be found with `getSessions`
 #'
-#' @param state US state abbreviation
+#' @param state US state abbreviation. Ignored when `sessionID` is supplied
 #'
 #' @param legiKey 32 character string provided by legiscan
 #'
@@ -28,7 +28,7 @@ getMasterListRaw <- function(sessionID = NULL, state = NULL, legiKey = NULL){
 
   response <- legiRequest(
     op = "getMasterListRaw",
-    state = state,
+    state = if (is.null(sessionID)) state else NULL,
     id = sessionID,
     legiKey = legiKey
   )